Map: Add and test IsEmpty method
