+++ builtin/v9/migration/test/util.go
@@ -5,15 +5,11 @@ import (
 	"crypto/sha256"
 	"errors"
 	"fmt"
-	"sync"
 	"testing"
 
 	"github.com/filecoin-project/go-state-types/actors"
 
 	"github.com/filecoin-project/go-state-types/rt"
-
-	block "github.com/ipfs/go-block-format"
-	ipldcbor "github.com/ipfs/go-ipld-cbor"
 
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/go-state-types/abi"
@@ -153,56 +149,6 @@ func initializeActor(ctx context.Context, t testing.TB, tree *builtin.ActorTree,
 	}
 	err = tree.SetActorV4(a, actor)
 	require.NoError(t, err)
-}
-
-type BlockStoreInMemory struct {
-	data map[cid.Cid]block.Block
-}
-
-func NewBlockStoreInMemory() *BlockStoreInMemory {
-	return &BlockStoreInMemory{make(map[cid.Cid]block.Block)}
-}
-
-func (mb *BlockStoreInMemory) Get(ctx context.Context, c cid.Cid) (block.Block, error) {
-	d, ok := mb.data[c]
-	if ok {
-		return d, nil
-	}
-	return nil, fmt.Errorf("not found")
-}
-
-func (mb *BlockStoreInMemory) Put(ctx context.Context, b block.Block) error {
-	mb.data[b.Cid()] = b
-	return nil
-}
-
-// Creates a new, empty IPLD store in memory.
-func NewADTStore(ctx context.Context) adt.Store {
-	return adt.WrapStore(ctx, ipldcbor.NewCborStore(NewBlockStoreInMemory()))
-
-}
-
-type SyncBlockStoreInMemory struct {
-	bs *BlockStoreInMemory
-	mu sync.Mutex
-}
-
-func NewSyncBlockStoreInMemory() *SyncBlockStoreInMemory {
-	return &SyncBlockStoreInMemory{
-		bs: NewBlockStoreInMemory(),
-	}
-}
-
-func (ss *SyncBlockStoreInMemory) Get(ctx context.Context, c cid.Cid) (block.Block, error) {
-	ss.mu.Lock()
-	defer ss.mu.Unlock()
-	return ss.bs.Get(ctx, c)
-}
-
-func (ss *SyncBlockStoreInMemory) Put(ctx context.Context, b block.Block) error {
-	ss.mu.Lock()
-	defer ss.mu.Unlock()
-	return ss.bs.Put(ctx, b)
 }
 
 type TestLogger struct {
